Issue #32. Add Posix error codes to ERRNOS enum.
Note that the current values for extended Posix error codes will be wrong
for all platforms, so they have been left as the (corrected) HAUSNUMERO
values. In the future, this could be handled by using a static class with
readonly int values instead of an enum; Mono.Unix.Native.Errno and
Mono.Unix.Native.NativeConvert.FromErrno can be used to set the correct
values on Mono, while the Windows values can be hard-coded.

- Set correct value for ENOTSUP (was HAUSNUMERO, should be HAUSNUMERO + 1)
- Add ENOTSOCK