Ignore non-deterministic monitor event specs.
Closes #103.

The address reported by ZMQ for Connected & Closed monitor events may be
garbage data and cannot be reliably tested. This is because the memory
address sent in the event message may point to a string that no longer
exists for whatever reason (socket closing, etc.).

This appears to be a known issue: LIBZMQ-450 (Jira).